update bridging services
